requirements.txt
langgraph
langchain-core
langchain-openai
pydantic
httpx
pytest
python-dotenv
google-search-results
beautifulsoup4
streamlit
pandas
ruff
